True initial commit
At this stage I consider the API itself good to go. There are a few skeletons here and there for features that are not yet implemented yet, such as automation for PRs and Issues. From a small sample round of testing and at least one true file committed and added via Git LFS, it seems we're ready to truly begin the project!